requirements.txt
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
pytesseract==0.3.10
pandas==2.2.1
requests==2.31.0
dill==0.3.7
ultralytics==8.1.29
pytest==8.1.1
flask==3.1.0
flask_cors==5.0.0
flasgger==0.9.7.1
python-dotenv==1.0.1
schedule==1.2.2
openai==1.59.7